Improved commentary and Exceptions. Moved functions to common
get_isomorphic_terms moved to commons.
